requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6
python-multipart==0.0.12
Jinja2==3.1.4
itsdangerous==2.2.0

# Database and ORM
sqlmodel==0.0.21
SQLAlchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
greenlet==3.1.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
cryptography==46.0.1

# HTTP Client
httpx==0.27.2
aiohttp==3.11.11
requests==2.32.3

# Data Validation
pydantic==2.10.5
pydantic_core==2.27.2
pydantic-settings==2.5.2

# Environment and Configuration
python-dotenv==1.0.1

# Async Support
anyio==4.6.2.post1
aiofiles==24.1.0

# Utilities
python-dateutil==2.9.0.post0
typing_extensions==4.12.2

# Google OAuth and Calendar Integration
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.153.0

# Development Tools (optional, can be removed for production)
watchfiles==1.1.1
